Extract default workspace test data into shared testing lib
Since these defaults are useful beyond just analysis, bringing them up a level makes them accessible to the whole codebase.
