+++ src/analysis/modals/CloudEnvironmentModal.test.ts
@@ -4,8 +4,6 @@ import { h, p } from 'react-hyperscript-helpers';
 import {
   azureDisk,
   azureRuntime,
-  defaultAzureWorkspace,
-  defaultGoogleWorkspace,
   defaultTestDisk,
   galaxyRunning,
   generateTestApp,
@@ -20,6 +18,7 @@ import { App } from 'src/libs/ajax/leonardo/models/app-models';
 import { Runtimes } from 'src/libs/ajax/leonardo/Runtimes';
 import { cloudProviderTypes } from 'src/libs/workspace-utils';
 import { asMockedFn } from 'src/testing/test-utils';
+import { defaultAzureWorkspace, defaultGoogleWorkspace } from 'src/testing/workspace-fixtures';
 
 type RuntimesAjaxExports = typeof import('src/libs/ajax/leonardo/Runtimes');
 type AppsAjaxExports = typeof import('src/libs/ajax/leonardo/Apps');
