+++ src/analysis/modals/HailBatchModal.test.ts
@@ -1,9 +1,10 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { h } from 'react-hyperscript-helpers';
-import { defaultAzureWorkspace, generateTestAppWithAzureWorkspace } from 'src/analysis/_testData/testData';
+import { generateTestAppWithAzureWorkspace } from 'src/analysis/_testData/testData';
 import { Apps } from 'src/libs/ajax/leonardo/Apps';
 import { asMockedFn } from 'src/testing/test-utils';
+import { defaultAzureWorkspace } from 'src/testing/workspace-fixtures';
 
 import { appToolLabels } from '../utils/tool-utils';
 import { HailBatchModal, HailBatchModalProps } from './HailBatchModal';
